librbd: enforce the underlying type of "Type"s

+++ src/tools/rbd_mirror/image_map/Types.cc
@@ -28,7 +28,7 @@ public:
 
   template <typename T>
   inline void operator()(const T& t) const {
-    ::encode(static_cast<uint32_t>(T::TYPE), m_bl);
+    ::encode(T::TYPE, m_bl);
     t.encode(m_bl);
   }
 private:
